Small refactor - move data classes out of the client
